- sdApi, translate and silero now asynchronous too - refactoring
